[frontend] Let the Service Worker update immediately
This means we won't need to reload the page a second time to activate the new service worker (and thus new assets which are precached).

Important: this only works well with the default GenerateSW workbox mode for Quasar. If we ever customise the Service Worker to other things (like Push) then we'd need to rethink this.